Move arg processing and log init to base class
